moved message to vga into output
