+++ src/main.rs
@@ -12,8 +12,8 @@ fn panic(info: &PanicInfo) -> ! {
 use bootloader::{entry_point, BootInfo};
 use popcorn::{
     hlt_loop, init,
-    low_level::vga_buffer::{clear_screen, Color, MessageToVga},
-    print_with_colors, println,
+    low_level::vga_buffer::{clear_screen, Color},
+    print_with_colors, println, userspace::output::MessageToVga,
 };
 entry_point!(kernel_main);
